refactor(types #37): exterminate the type-name alias tables — full passthrough
Frame has no type system: type names pass through to generated code
VERBATIM (docs/frame_language.md). The per-backend type-alias tables that
translated int→i64 / str→String / Any→object / … contradicted that
contract and were the lone reason `list`/`dict` "didn't work". Removed
across all 8 statically-typed backends:

- 7 dogfooded mappers (rust/cpp/csharp/java/kotlin/go/swift *.frs +
  regenerated *.gen.rs) → passthrough.
- C's two inline sites (backends/c.rs convert_type_to_c, interface_gen.rs
  return-type) → passthrough.

Kept ONLY genuinely-structural transforms (not name-aliasing): Rust
borrow→owned widening (&str→String, &[T]→Vec<T>; load-bearing for owned
storage + framec's own dogfooded FSMs), Kotlin void→Unit, Go void→empty,
Swift void→Void + nullable/array syntax, C's runtime-container ABI
(list→FrameVec*, dict→FrameDict*). framec self-hosts; lib tests 473/0.

Snapshot corpus converted to per-backend native fixtures: since one shared
Rust-flavored fixture can no longer render natively for every backend, each
backend gets fixtures written in its own type names. tests/fixtures/
_canonical/ templates + scripts/gen_snapshot_fixtures.py generate
tests/fixtures/<target>/ (17 backends x 13). compile_fixture loads the
per-target dir. Reblessing fixed latent leaks (e.g. C++ domain fields
i32/String → int/std::string). All 17 snapshot suites green.

Note: RFC-0034 compile-checks + the Docker matrix (real per-language
compilation) require the language toolchains and must run in CI.

+++ framec/tests/rust_snapshots.rs
@@ -654,10 +654,10 @@ fn bug31_no_std_portable_paths() {
     let src = r#"
 @@system Foo {
     interface:
-        bar(): int
+        bar(): i64
     machine:
         $A {
-            bar(): int { @@:(1) }
+            bar(): i64 { @@:(1) }
         }
 }
 "#;
@@ -758,14 +758,14 @@ fn bug33_generated_compiles_no_std() {
 @@system Bug33 {
     interface:
         bump()
-        total(): int
+        total(): i64
     machine:
         $A {
             bump()       { -> $B }
-            total(): int { @@:(0) }
+            total(): i64 { @@:(0) }
         }
         $B {
-            total(): int { @@:(1) }
+            total(): i64 { @@:(1) }
         }
 }
 "#,
@@ -835,21 +835,21 @@ fn bug34_enter_args_type_faithful() {
 @@system Bug34 {
     interface:
         run()
-        sum(): int
-        head(): int
+        sum(): i64
+        head(): i64
     machine:
         $Start { run() { -> (42) $Mid } }
         $Mid {
-            $>(n: int)   { self.total = n + 1 }
-            sum(): int   { @@:(self.total) }
+            $>(n: i64)   { self.total = n + 1 }
+            sum(): i64   { @@:(self.total) }
             run()        { -> (vec![99]) $End }
         }
         $End {
             $>(xs: Vec<i64>) { self.total = xs[0] }
-            head(): int      { @@:(self.total) }
+            head(): i64      { @@:(self.total) }
         }
     domain:
-        total: int = 0
+        total: i64 = 0
 }
 "#,
         "rust",
@@ -939,17 +939,17 @@ fn bug35_start_state_exit_arg_binds() {
 @@system Bug35 {
     interface:
         go()
-        seen(): int
+        seen(): i64
     machine:
         $A {
             go()          { (99) -> $B }
-            <$(code: int) { self.s = code }
+            <$(code: i64) { self.s = code }
         }
         $B {
-            seen(): int { @@:(self.s) }
+            seen(): i64 { @@:(self.s) }
         }
     domain:
-        s: int = 0
+        s: i64 = 0
 }
 "#,
         "rust",
@@ -1025,23 +1025,23 @@ fn pop_decorated_args_typed() {
         go()
         nest()
         back()
-        entered(): int
-        exited(): int
+        entered(): i64
+        exited(): i64
     machine:
         $Start { go() { -> $Idle } }
         $Idle {
-            $>(tag: int)   { self.e = self.e + tag }
+            $>(tag: i64)   { self.e = self.e + tag }
             nest()         { push$ -> $Nested }
-            entered(): int { @@:(self.e) }
-            exited(): int  { @@:(self.x) }
+            entered(): i64 { @@:(self.e) }
+            exited(): i64  { @@:(self.x) }
         }
         $Nested {
-            <$(code: int) { self.x = code }
+            <$(code: i64) { self.x = code }
             back()        { (7) -> (42) pop$ }
         }
     domain:
-        e: int = 0
-        x: int = 0
+        e: i64 = 0
+        x: i64 = 0
 }
 "#,
         "rust",
